+++ Core/libft/ft_strchr.c
@@ -6,7 +6,7 @@
 /*   By: lfrench <lfrench@student.42luxembourg      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/23 14:00:11 by lfrench           #+#    #+#             */
-/*   Updated: 2024/03/02 20:28:11 by lfrench          ###   ########.fr       */
+/*   Updated: 2024/03/02 21:18:14 by lfrench          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 char	*ft_strchr(const char *s, int c)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i] != '\0')
